Improve release-check termination
cfncluster-release-check integration test runs for few hours spawning
tens of clusters across the supported regions.
Interrupting this process means to leave a number of clusters up and
running.

This patch wants to gracefully manage the interrupted termination,
cleaning all the clusters that have been created.
This is obtained with the following changes
1. Capture the terminate/interrupt signals and manage the termination of
the running processes. This will raise an exception in all those threads
where the main operations are running and prevents further commands to
be executed. The finally block then copes with the removal of the
created clusters.
2. The name of the cluster has been made reasonably unique by adding
the timestamp of the launch. In this way we have not to wait all the
clusters to be deleted and we can launch a second instance.
3. The delete cluster operation is now not suspensive.